feat: `cargo-pgx init` does initdb
* Migrate `cargo-pgx::commands::run::initdb` to `cargo-pgx::commands::init::initdb`.
* Call `cargo-pgx::commands::init::iinitdb` in `cargo-pgx pgx init`.
* Include backwards compatibility safety, still call `initdb` in `cargo-pgx pgx run`.
* Updates `README.md`.
